Extract processos tab into dedicated screen
